\name{barah-package}
\alias{barah-package}
\alias{barah}
\docType{package}
\title{
  Format elements into a specified pattern
}
\description{
  Data comes in messy, unpredictable formats. Use the fashion function to bring 
  order to the chaos.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
David Ranzolin

Maintainer: David Ranzolin <daranzolin@gmail.com>
}
\keyword{ package }
