% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fashion.R
\name{fashion}
\alias{fashion}
\title{Fashion an element according to a specified pattern}
\usage{
fashion(x, pattern, replace_char = "X", ignore = NULL)
}
\arguments{
\item{x}{an atomic vector}

\item{pattern}{specified pattern with pattern_replace characters}

\item{replace_char}{character specifying replacement}

\item{ignore}{characters to ignore. Multiple characters should be separated by regex vertical pipe operator |}
}
\value{
character vector
}
\description{
Fashion an element according to a specified pattern
}
\examples{
fashion(c("1234567890", "0987654321"), "(XXX) XXX-XXXX")
}
